@machawk1 2018-07-09
@ibnesayeed Why were the rates with threading so much lower? I would expect them to be higher on average.