@machawk1 2018-07-14
> observed some really strange behaviors

Please document them here.

> isolate certain things in a separate script and test them.

Which things? As we talked about in-person, the Py threading mechanism ought to be compared to the base but that is likely not the culprit.