@ibnesayeed 2018-07-16
Created a separate benchmarking repository ([IPFS API Concurrency Test](https://github.com/ibnesayeed/ipfsapi-concurrency-test)) and reported the observation in the API repo (https://github.com/ipfs/py-ipfs-api/issues/131).