@funkyfuture 2018-08-19
maybe i'm missing a point, but have you considered to put a wsgi daemon in front of ipwb?