@funkyfuture 2018-08-19
> I am not sure how a web server used has anything to do with how the content will be read from the IPFS server.

not at all.

from your benchmark i had the impression that the aim is to improve response time to web requests. a threading wsgi host process would facilitate that.

if you are onto reading header and payload concurrently, [trio](https://trio.readthedocs.io/) is an option for Python 3.5+. (i know, but maybe that would motivate the port to Python 3.)